config apollo client graphql
