+++ client/components/authenticate/index.js
@@ -1,1 +1,1 @@
-export { LoginScreen } from './login';
+export { SignUpScreen } from './signup';
