--- client/components/authenticate/login.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import { View } from 'react-native';
-
-import { CommonInput, CommonButton, LoadingSpine } from '../common';
-
-const styles = {
-  styleContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'white',
-  },
-};
-
-export const LoginScreen = () => (
-  <View style={styles.styleContainer}>
-    <CommonInput style={styles} />
-    <CommonButton>Hello</CommonButton>
-    <LoadingSpine sourceImg={require('../../styles/img/loadingSpine.png')} />
-  </View>
-);
