--- client/sagas/rootSaga.js
@@ -1,5 +1,0 @@
-import { all } from 'redux-saga/effects';
-
-export default function*() {
-  yield all([]);
-}
