--- client/store/index.js
@@ -1,16 +1,0 @@
-import { compose, createStore, applyMiddleware } from 'redux';
-import createSageMiddleWare from 'redux-saga';
-
-import rootReducer from '../reducers/rootReducer';
-import rootSaga from '../sagas/rootSaga';
-
-const sagaMiddleware = createSageMiddleWare();
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-export const store = createStore(
-  rootReducer,
-  composeEnhancers(applyMiddleware(sagaMiddleware))
-);
-
-sagaMiddleware.run(rootSaga);
