+++ package.json
@@ -12,6 +12,10 @@
   "dependencies": {
     "@babel/core": "^7.4.4",
     "@babel/preset-env": "^7.4.4",
+    "apollo-boost": "^0.4.0",
+    "apollo-cache-inmemory": "^1.6.0",
+    "apollo-client": "^2.6.0",
+    "apollo-link-http": "^1.5.14",
     "babel-loader": "^8.0.6",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
@@ -32,6 +36,7 @@
     "nodemon": "^1.19.0",
     "prop-types": "^15.7.2",
     "react": "16.5.0",
+    "react-apollo": "^2.5.6",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
     "react-navigation": "^3.11.0",
     "react-redux": "^7.0.3",
